moved render_label from pp to pl and renamed it to render_labels
