+++ spatial_data/pl/__init__.py
@@ -1,6 +1,6 @@
-from .utils import _get_linear_colormap, _get_listed_colormap
 from .plot import PlotAccessor
 from .spectra import plot_expression_spectra
+from .utils import _get_linear_colormap, _get_listed_colormap
 
 __all__ = [
     "PlotAccessor",
