+++ spatial_data/pl/utils.py
@@ -1,59 +1,14 @@
-from typing import List, Union
+from typing import List
 
-from matplotlib.colors import LinearSegmentedColormap, ListedColormap
 import numpy as np
+from matplotlib.colors import LinearSegmentedColormap, ListedColormap
 from skimage.segmentation import find_boundaries
+
+from ..pp.utils import _normalize
+
 
 def _get_linear_colormap(colors: list, background: str):
     return [LinearSegmentedColormap.from_list(c, [background, c], N=256) for c in colors]
-
-
-def _get_listed_colormap(color_dict: dict):
-    sorted_labels = sorted(color_dict.keys())
-    colors = [color_dict[k] for k in sorted_labels]
-    cmap = ListedColormap(["black"] + colors, N=len(colors) + 1)
-    return cmap
-
-
-
-def _normalize(
-    img: np.ndarray,
-    pmin: float = 3.0,
-    pmax: float = 99.8,
-    eps: float = 1e-20,
-    clip: bool = False,
-    name: str = "normed",
-) -> np.ndarray:
-    """Performs a min max normalisation.
-
-    This function was adapted from the csbdeep package.
-
-    Parameters
-    ----------
-    dataarray: xr.DataArray
-        A xarray DataArray with an image field.
-    pmin: float
-        Lower quantile (min value) used to perform qunatile normalization.
-    pmax: float
-        Upper quantile (max value) used to perform qunatile normalization.
-    eps: float
-        Epsilon float added to prevent 0 division.
-    clip: bool
-        Ensures that normed image array contains no values greater than 1.
-
-    Returns
-    -------
-    xr.DataArray
-        A min-max normalized image.
-    """
-    perc = np.percentile(img, [pmin, pmax], axis=(1, 2)).T
-
-    norm = (img - np.expand_dims(perc[:, 0], (1, 2))) / (np.expand_dims(perc[:, 1] - perc[:, 0], (1, 2)) + eps)
-
-    if clip:
-        norm = np.clip(norm, 0, 1)
-
-    return norm
 
 
 def _colorize(
@@ -97,7 +52,7 @@ def _colorize(
     return colored
 
 
-def _render_label(mask, cmap_mask, img=None, alpha=0.2, alpha_boundary=1.0, mode="inner"):
+def _render_labels(mask, cmap_mask, img=None, alpha=0.2, alpha_boundary=1.0, mode="inner"):
     colored_mask = cmap_mask(mask)
 
     mask_bool = mask > 0
@@ -114,3 +69,29 @@ def _render_label(mask, cmap_mask, img=None, alpha=0.2, alpha_boundary=1.0, mode
     im[mask_bound] = alpha_boundary * colored_mask[mask_bound] + (1 - alpha_boundary) * img[mask_bound]
 
     return im
+
+
+def _get_listed_colormap(color_dict: dict):
+    sorted_labels = sorted(color_dict.keys())
+    colors = [color_dict[k] for k in sorted_labels]
+    cmap = ListedColormap(["black"] + colors, N=len(colors) + 1)
+    return cmap
+
+
+def _label_segmentation_mask(segmentation: np.ndarray, annotations: dict) -> np.ndarray:
+    """
+    Relabels a segmentation according to the annotations df (contains the columns type, cell).
+    """
+    labeled_segmentation = segmentation.copy()
+    all_cells = []
+
+    for k, v in annotations.items():
+        mask = np.isin(segmentation, v)
+        labeled_segmentation[mask] = k
+        all_cells.extend(v)
+
+    # remove cells that are not indexed
+    neg_mask = ~np.isin(segmentation, all_cells)
+    labeled_segmentation[neg_mask] = 0
+
+    return labeled_segmentation
