+++ spatial_data/pp/preprocessing.py
@@ -11,16 +11,10 @@ from skimage.restoration import unsupervised_wiener
 from skimage.segmentation import expand_labels
 
 from ..base_logger import logger
-from ..constants import COLORS, Attrs, Dims, Features, Layers, Props
+from ..constants import COLORS, Dims, Features, Layers, Props
 from ..la.label import _format_labels
-from ..pl import _get_listed_colormap
 from .intensity import sum_intensity
-from .utils import (
-    _label_segmentation_mask,
-    _relabel_cells,
-    _remove_unlabeled_cells,
-    _render_label,
-)
+from .utils import _normalize, _relabel_cells, _remove_unlabeled_cells
 
 
 @xr.register_dataset_accessor("pp")
@@ -820,84 +814,3 @@ class PreprocessingAccessor:
 
         # adding the default obs back to the object
         return obj.pp.add_observations()
-
-
-    def render_label(
-        self, alpha: float = 0, alpha_boundary: float = 1, mode: str = "inner", override_color: Union[str, None] = None
-    ) -> xr.Dataset:
-        """
-        Render the labeled cells in the data object.
-
-        This method renders the labeled cells in the data object based on the label colors and segmentation.
-        The rendered visualization is represented in RGBA format.
-
-        Parameters
-        ----------
-        alpha : float, optional
-            The alpha value to control the transparency of the rendered labels. Default is 0.
-        alpha_boundary : float, optional
-            The alpha value for boundary pixels in the rendered labels. Default is 1.
-        mode : str, optional
-            The mode for rendering the labels: "inner" for internal region, "boundary" for boundary pixels.
-            Default is "inner".
-        override_color : any, optional
-            The color value to override the default label colors. Default is None.
-
-        Returns
-        -------
-        any
-            The updated data object with the rendered labeled cells as a new plot layer.
-
-        Raises
-        ------
-        AssertionError
-            If the data object does not contain label information. Use 'add_labels' function to add labels first.
-
-        Notes
-        -----
-        - The function retrieves label colors from the data object and applies the specified alpha values and mode.
-        - It renders the labeled cells based on the label colors and the segmentation layer.
-        - The rendered visualization is represented in RGBA format and added as a new plot layer to the data object.
-        - If 'override_color' is provided, all labels will be rendered using the specified color.
-        """
-        assert Layers.LABELS in self._obj, "Add labels via the add_labels function first."
-
-        # TODO: Attribute class in constants.py
-        color_dict = self._obj.la._label_to_dict(Props.COLOR, relabel=True)
-        if override_color is not None:
-            color_dict = {k: override_color for k in color_dict.keys()}
-
-        cmap = _get_listed_colormap(color_dict)
-
-        cells_dict = self._obj.la._cells_to_label(relabel=True)
-        segmentation = self._obj[Layers.SEGMENTATION].values
-        mask = _label_segmentation_mask(segmentation, cells_dict)
-
-        if Layers.PLOT in self._obj:
-            attrs = self._obj[Layers.PLOT].attrs
-            rendered = _render_label(
-                mask,
-                cmap,
-                self._obj[Layers.PLOT].values,
-                alpha=alpha,
-                alpha_boundary=alpha_boundary,
-                mode=mode,
-            )
-            self._obj = self._obj.drop_vars(Layers.PLOT)
-        else:
-            attrs = {}
-            rendered = _render_label(mask, cmap, alpha=alpha, alpha_boundary=alpha_boundary, mode=mode)
-
-        da = xr.DataArray(
-            rendered,
-            coords=[
-                self._obj.coords[Dims.Y],
-                self._obj.coords[Dims.X],
-                ["r", "g", "b", "a"],
-            ],
-            dims=[Dims.Y, Dims.X, Dims.RGBA],
-            name=Layers.PLOT,
-            attrs=attrs,
-        )
-
-        return xr.merge([self._obj, da])
